Use int for mti and mri

+++ gtk2_ardour/tracker/grid_header.h
@@ -42,7 +42,7 @@ public:
 	void setup_track_headers ();
 
 	void set_time_header_size (int width=-1, int height=-1);
-	void set_track_header_size (size_t mti, int width=-1, int height=-1);
+	void set_track_header_size (int mti, int width=-1, int height=-1);
 
 	/**
 	 * Align the time and track headers with the time and track columns.
